[REFACTOR] Login and register moved to users package
