+++ Eventy/app/src/main/java/com/example/eventy/services/ServiceCreationFragment.java
@@ -18,7 +18,7 @@ import android.view.View;
 import android.view.ViewGroup;
 import com.example.eventy.R;
 import com.example.eventy.databinding.FragmentServiceCreationBinding;
-import com.example.eventy.register.CarouselAdapter;
+import com.example.eventy.users.register.CarouselAdapter;
 
 import java.util.ArrayList;
 import java.util.Arrays;
