+++ Eventy/app/src/main/java/com/example/eventy/users/register/CarouselAdapter.java
@@ -1,4 +1,4 @@
-package com.example.eventy.register;
+package com.example.eventy.users.register;
 
 import android.net.Uri;
 import android.view.LayoutInflater;
