+++ Eventy/app/src/main/res/layout/fragment_login.xml
@@ -9,7 +9,7 @@
     android:gravity="center"
     android:padding="40dp"
     android:background="@color/background_color"
-    tools:context="login.LoginFragment">
+    tools:context=".users.login.LoginFragment">
 
     <!-- Login Label -->
     <TextView
