+++ Eventy/app/src/main/res/layout/fragment_register.xml
@@ -13,7 +13,7 @@
         android:gravity="center"
         android:padding="40dp"
         android:background="@color/background_color"
-        tools:context="register.RegisterFragment">
+        tools:context=".users.register.RegisterFragment">
         <TextView
             android:id="@+id/register_label"
             android:layout_width="wrap_content"
