+++ Eventy/app/src/main/res/layout/fragment_register_provider.xml
@@ -2,7 +2,7 @@
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
 xmlns:app="http://schemas.android.com/apk/res-auto"
 xmlns:tools="http://schemas.android.com/tools"
-tools:context=".register.RegisterProviderFragment"
+tools:context=".users.register.RegisterProviderFragment"
 android:layout_width="match_parent"
 android:layout_height="match_parent"
 android:orientation="vertical"
