+++ Eventy/app/src/main/res/layout/fragment_service_creation.xml
@@ -12,7 +12,7 @@
         android:gravity="center"
         android:padding="40dp"
         android:background="@color/background_color"
-        tools:context="register.RegisterFragment">
+        tools:context=".users.register.RegisterFragment">
 
         <TextView
             android:id="@+id/add_service_title"
