+++ Eventy/app/src/main/res/layout/fragment_user_fast_registration_upgrade_account_provider.xml
@@ -4,7 +4,7 @@
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".register.RegisterOrganiserFragment"
+    tools:context=".users.register.RegisterOrganiserFragment"
     android:layout_width="match_parent"
     android:fillViewport="true"
     android:background="@color/background_color"
@@ -12,7 +12,7 @@
     android:gravity="center">
 
     <LinearLayout
-    tools:context=".register.RegisterProviderFragment"
+    tools:context=".users.register.RegisterProviderFragment"
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     android:orientation="vertical"
