+++ Eventy/app/src/main/res/navigation/mobile_navigation.xml
@@ -13,13 +13,13 @@
 
     <fragment
         android:id="@+id/nav_login"
-        android:name="com.example.eventy.login.LoginFragment"
+        android:name="com.example.eventy.users.login.LoginFragment"
         android:label="Login"
         tools:layout="@layout/fragment_login" />
 
     <fragment
         android:id="@+id/nav_register"
-        android:name="com.example.eventy.register.RegisterFragment"
+        android:name="com.example.eventy.users.register.RegisterFragment"
         android:label="Register"
         tools:layout="@layout/fragment_register" />
